+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Added `should-wait = true` to the EC2 instance configuration so it doesn't register itself in the ECS cluster while in warm pool.
+- Fixed WarmPool usage in the EC2 Auto Scaling Group.
 
 ## [0.2.1] - 2024-02-28
 ### Fixed